requirements.txt
Flask
opencv-python
numpy
Pillow
fastdeploy-python